Small heap of fixes.
We actually clean up web request responses now. That means the feeds are much less clogged.

We pay attention to the configured minimum interval for checking feeds. That means we don't check
each feed once per minute.

We look for log4net.config in the right place. Previously, you had to put it in /, because that's
the working directory for the mod_mono process. Now we detect the application folder and use that.
That was so problematic when I couldn't figure out why things weren't logging...