Improve date parsing; handle bodiless articles
ISO8601 is a family of formats; '2017-01-15' is proper ISO8601, as is
'2017-01-15T15:43:12.00013415+1030'. I extended the parsing and
formatting code to handle multiformat bundles.

Articles without bodies caused segfaults. (The memory handler
performed admirably, so I was able to catch these problems and the
program didn't abort.) I fixed this.

Also included a little more info about logfiles in the readme.

Fixes #6