Keep the same reference for feeds / subscriptions
On the JS side, we're getting new objects for feeds and subscriptions
every time we check. However, we keep old articles and don't update
them to the new feeds. This means we don't update feed read/unread
counts correctly.

By holding onto the old objects and updating them with the new fields,
we get updates and the unread counts work.